stores: remove contract set store
