+++ internal/stores/contracts_test.go
@@ -439,3 +439,59 @@ func TestAncestorsContracts(t *testing.T) {
 		}
 	}
 }
+
+func (s *SQLStore) addTestContract(fcid types.FileContractID, hk consensus.PublicKey) (bus.Contract, error) {
+	rev := testContractRevision(fcid, hk)
+	return s.AddContract(rev, types.ZeroCurrency, 0)
+}
+
+func (s *SQLStore) addTestRenewedContract(fcid, renewedFrom types.FileContractID, hk consensus.PublicKey, startHeight uint64) (bus.Contract, error) {
+	rev := testContractRevision(fcid, hk)
+	return s.AddRenewedContract(rev, types.ZeroCurrency, startHeight, renewedFrom)
+}
+
+func testContractRevision(fcid types.FileContractID, hk consensus.PublicKey) rhpv2.ContractRevision {
+	uc, _ := types.GenerateDeterministicMultisig(1, 2, "salt")
+	uc.PublicKeys[1].Key = hk[:]
+	uc.Timelock = 192837
+	return rhpv2.ContractRevision{
+		Revision: types.FileContractRevision{
+			ParentID:          fcid,
+			UnlockConditions:  uc,
+			NewRevisionNumber: 200,
+			NewFileSize:       4096,
+			NewFileMerkleRoot: crypto.Hash{222},
+			NewWindowStart:    400,
+			NewWindowEnd:      500,
+			NewValidProofOutputs: []types.SiacoinOutput{
+				{
+					Value:      types.NewCurrency64(121),
+					UnlockHash: types.UnlockHash{2, 1, 2},
+				},
+			},
+			NewMissedProofOutputs: []types.SiacoinOutput{
+				{
+					Value:      types.NewCurrency64(323),
+					UnlockHash: types.UnlockHash{2, 3, 2},
+				},
+			},
+			NewUnlockHash: types.UnlockHash{6, 6, 6},
+		},
+		Signatures: [2]types.TransactionSignature{
+			{
+				ParentID:       crypto.Hash(fcid),
+				PublicKeyIndex: 0,
+				Timelock:       100000,
+				CoveredFields:  types.FullCoveredFields,
+				Signature:      []byte("signature1"),
+			},
+			{
+				ParentID:       crypto.Hash(fcid),
+				PublicKeyIndex: 1,
+				Timelock:       200000,
+				CoveredFields:  types.FullCoveredFields,
+				Signature:      []byte("signature2"),
+			},
+		},
+	}
+}
